requirements.txt
click==7.1.2
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.3
opencv-python==4.5.2.52
Pillow==8.2.0
py-midicsv==2.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytube==10.8.2
six==1.16.0
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111
tqdm==4.60.0
typing-extensions==3.10.0.0
